hopefully the last linter cleanup

+++ libbeat/metric/system/filesystem/filesystem_unix.go
@@ -35,11 +35,11 @@ func (fs *FSStat) GetUsage() error {
 		return fmt.Errorf("error in Statfs syscall: %w", err)
 	}
 
-	fs.Total = opt.UintWith(uint64(stat.Blocks)).MultUint64OrNone(uint64(stat.Bsize))
-	fs.Free = opt.UintWith(uint64(stat.Bfree)).MultUint64OrNone(uint64(stat.Bsize))
-	fs.Avail = opt.UintWith(uint64(stat.Bavail)).MultUint64OrNone(uint64(stat.Bsize))
+	fs.Total = opt.UintWith(stat.Blocks).MultUint64OrNone(uint64(stat.Bsize))
+	fs.Free = opt.UintWith(stat.Bfree).MultUint64OrNone(uint64(stat.Bsize))
+	fs.Avail = opt.UintWith(stat.Bavail).MultUint64OrNone(uint64(stat.Bsize))
 	fs.Files = opt.UintWith(stat.Files)
-	fs.FreeFiles = opt.UintWith(uint64(stat.Ffree))
+	fs.FreeFiles = opt.UintWith(stat.Ffree)
 
 	fs.fillMetrics()
